agent/requirements.txt
fastapi~=0.115.0
httpx~=0.27.0
uvicorn[standard]~=0.34.0
pydantic-settings~=2.7.0
python-dotenv~=1.0.0
langchain-google-genai
langchain-core
langchain-mcp-adapters>=0.2.2
langgraph
structlog~=24.4.0
sse-starlette~=2.2.0
temporalio~=1.6.0
guardrails-ai
slowapi~=0.1.9
# Specific hub validators will be installed via CLI or extra requirements
